Worked on Playlist & User

+++ server/src/main/java/com/transteven/multiplaylist/playlist/Playlist.java
@@ -1,6 +1,6 @@
 package com.transteven.multiplaylist.playlist;
 
-import com.transteven.multiplaylist.playlist.PlaylistDTO;
+import com.transteven.multiplaylist.playlist.dto.PlaylistDTO;
 import com.transteven.multiplaylist.video.Video;
 import java.util.List;
 import javax.persistence.Column;
@@ -9,7 +9,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
-import javax.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -23,7 +22,7 @@ public class Playlist {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(insertable = false, updatable = false)
-  private Long id;
+  private int id;
 
   @Column(columnDefinition = "VARCHAR(255)", nullable = false)
   private String name;
@@ -31,7 +30,7 @@ public class Playlist {
   @OneToMany(mappedBy = "playlist")
   private List<Video> playlistVideos;
 
-  public Playlist(Long id, String name) {
+  public Playlist(int id, String name) {
     this.id = id;
     this.name = name;
   }
