+++ server/src/main/java/com/transteven/multiplaylist/playlist/PlaylistRepository.java
@@ -2,5 +2,6 @@ package com.transteven.multiplaylist.playlist;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 @Repository
-public interface PlaylistRepository extends JpaRepository<Playlist, Long>{}
+public interface PlaylistRepository extends JpaRepository<Playlist, Integer> {}
